3_backend_python/requirements.txt
flask
flask_cors
google-generativeai
python-dotenv
numpy
librosa
pydub
soundfile